backend/pyproject.toml
[project]
name = "atlas-haven-backend"
version = "0.1.0"
description = "Atlas Haven backend control plane"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.14.0",
  "arq>=0.27.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.8.0",
  "python-multipart>=0.0.20",
  "pymysql>=1.1.1",
  "pyyaml>=6.0.2",
  "redis>=5.2.0,<6.0.0",
  "sqlalchemy>=2.0.40",
  "structlog>=25.0.0",
  "uvicorn>=0.35.0",
  "aura-memory>=0.1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=1.0.0",
  "ruff>=0.12.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
  "unit: Unit tests (no HTTP client, no DB)",
  "integration: Integration tests (full HTTP round-trip via TestClient)",
  "contract: OpenAPI contract verification",
  "platform: Platform core — health, services, updates, settings",
  "chat: Chat sessions and messages",
  "cognitive: Cognitive memory (AuraSDK)",
  "ollama: Local LLM / Ollama endpoints",
  "benchmark: Benchmark run, results, submission",
  "knowledge_base: RAG document pipeline",
  "maps: Offline map management",
  "zim: ZIM / Wikipedia content",
  "easy_setup: Guided setup wizard",
  "content_updates: Manifest and content updates",
  "downloads: Download job tracking",
  "docs: Documentation listing",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
